# align with RFC 3987 IRIs