Fixes #58

The performance impact appears to be minor, but if it is a concern then I think the regular expression repetitions can be optimized.

**master**
```
goos: linux
goarch: amd64
pkg: mvdan.cc/xurls/v2
cpu: Intel(R) Core(TM) i5-4460  CPU @ 3.20GHz
BenchmarkStrictEmpty-4     	69208698	        14.87 ns/op
BenchmarkStrictSingle-4    	   30261	     39343 ns/op
BenchmarkStrictMany-4      	    9802	    121895 ns/op
BenchmarkRelaxedEmpty-4    	   75169	     14815 ns/op
BenchmarkRelaxedSingle-4   	   12896	     91612 ns/op
BenchmarkRelaxedMany-4     	    4803	    242835 ns/op
```

**this PR**
```
goos: linux
goarch: amd64
pkg: mvdan.cc/xurls/v2
cpu: Intel(R) Core(TM) i5-4460  CPU @ 3.20GHz
BenchmarkStrictEmpty-4     	76394869	        14.82 ns/op
BenchmarkStrictSingle-4    	   30200	     39027 ns/op
BenchmarkStrictMany-4      	    9250	    122698 ns/op
BenchmarkRelaxedEmpty-4    	   78207	     14835 ns/op
BenchmarkRelaxedSingle-4   	   13330	     92321 ns/op
BenchmarkRelaxedMany-4     	    5029	    238684 ns/op
```