@mvdan 2022-01-19
Also, could you please squash the commits and include benchmark numbers in the commit message? Ideally with at least 6 runs before and after, and using https://pkg.go.dev/golang.org/x/perf/cmd/benchstat to compare the results. The regex being a few percent slower is not a worry as long as it becomes more correct.